refactor: create unwrap pk and sig functions
